fix type issues

+++ romancal/tweakreg/tweakreg_step.py
@@ -68,7 +68,7 @@ class TweakRegStep(RomanStep):
         output_use_model = boolean(default=True)  # When saving use `DataModel.meta.filename`
         update_source_catalog_coordinates = boolean(default=False) # Update source catalog file with tweaked coordinates?
         save_l1_wcs = boolean(default=True)
-        vo_timeout = integer(min=0, default=120) # VO catalog service timeout.
+        vo_timeout = float(min=0, default=120.) # VO catalog service timeout.
     """
 
     reference_file_types: ClassVar = []
